requirements.txt
streamlit>=1.35.0
openai>=1.30.0
requests>=2.31.0
faster-whisper>=1.0.0
python-docx>=1.1.0
